refactor: reorganize files into feature modules
