+++ src/features/auth/context/AuthContext.jsx
@@ -1,5 +1,5 @@
 import { createContext, useEffect, useState } from 'react'
-import { supabase, isSupabaseConfigured } from '../supabaseClient'
+import { supabase, isSupabaseConfigured } from '@/supabaseClient'
 
 // eslint-disable-next-line react-refresh/only-export-components
 export const AuthContext = createContext({ user: null, role: null })
