+++ src/features/auth/hooks/useAccount.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabaseClient'
+import { supabase } from '@/supabaseClient'
 
 export function useAccount() {
   const updateProfile = (data) => supabase.auth.updateUser({ data })
