+++ src/features/auth/hooks/useSupabaseAuth.js
@@ -1,6 +1,6 @@
 import { useState } from 'react'
-import { supabase } from '../supabaseClient'
-import { pushNotification } from '../utils/notifications'
+import { supabase } from '@/supabaseClient'
+import { pushNotification } from '@/utils/notifications'
 
 export function useSupabaseAuth() {
   const [error, setError] = useState(null)
