+++ src/features/chat/components/AttachmentPreview.jsx
