+++ src/features/dashboard/components/AuditTrail.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { supabase } from '../supabaseClient'
+import { supabase } from '@/supabaseClient'
 import Spinner from './Spinner'
 import ErrorMessage from './ErrorMessage'
