+++ src/features/dashboard/components/InventoryTabs.jsx
@@ -3,22 +3,22 @@ import React, { useState, useEffect } from 'react'
 import { useForm } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { supabase } from '../supabaseClient'
-import { handleSupabaseError } from '../utils/handleSupabaseError'
+import { supabase } from '@/supabaseClient'
+import { handleSupabaseError } from '@/utils/handleSupabaseError'
 import HardwareCard from './HardwareCard'
 import TaskCard from './TaskCard'
-import ChatTab from './ChatTab'
+import ChatTab from '@/features/chat/components/ChatTab'
 import { PlusIcon, ChatBubbleOvalLeftIcon } from '@heroicons/react/24/outline'
-import { linkifyText } from '../utils/linkify'
+import { linkifyText } from '@/utils/linkify'
 import { toast } from 'react-hot-toast'
 import ConfirmModal from './ConfirmModal'
 import Spinner from './Spinner'
 import ErrorMessage from './ErrorMessage'
 import { useHardware } from '../hooks/useHardware'
 import { useTasks } from '../hooks/useTasks'
-import { useChatMessages } from '../hooks/useChatMessages'
+import { useChatMessages } from '@/features/chat/hooks/useChatMessages'
 import { useObjects } from '../hooks/useObjects'
-import { useAuth } from '../hooks/useAuth'
+import { useAuth } from '@/features/auth/hooks/useAuth'
 import { useNavigate } from 'react-router-dom'
 
 const TAB_KEY = (objectId) => `tab_${objectId}`
