+++ src/features/dashboard/components/TaskCard.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Card from './Card'
+import Card from '@/components/Card'
 import { PencilIcon, TrashIcon } from '@heroicons/react/24/outline'
 
 /**
