+++ src/features/dashboard/hooks/useHardware.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabaseClient'
+import { supabase } from '@/supabaseClient'
 
 export function useHardware() {
   const fetchHardware = (objectId) =>
