+++ src/features/dashboard/hooks/useObjects.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabaseClient'
+import { supabase } from '@/supabaseClient'
 
 export function useObjects() {
   const fetchObjects = () =>
