+++ src/features/dashboard/hooks/useTasks.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabaseClient'
+import { supabase } from '@/supabaseClient'
 
 export function useTasks() {
   const fetchTasks = (objectId) =>
