+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App'
 import './index.css'
-import { AuthProvider } from './context/AuthContext'
+import { AuthProvider } from '@/features/auth/context/AuthContext'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
