+++ src/utils/handleSupabaseError.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabaseClient'
+import { supabase } from '@/supabaseClient'
 import { toast } from 'react-hot-toast'
 
 export async function handleSupabaseError(error, navigate, message) {
