+++ vite.config.js
@@ -11,6 +11,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, 'src'),
+      '@features': path.resolve(__dirname, 'src/features'),
     },
   },
   server: {
